Fix exit status of Makefile's lint recipe
make lint returns 0 exit code if the last line succeed regardless
of the exit code of previous lines.

If the .ONESHELL special target appears anywhere in the makefile
then all recipe lines for each target will be provided to a single
invocation of the shell. Source:
https://www.gnu.org/software/make/manual/html_node/One-Shell.html

It's currently not possible to specify .ONESHELL only for individual
targets.